+++ src/main/java/com/example/employee/model/entity/Customer.java
@@ -10,7 +10,7 @@ public class Customer {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer customerId;
+    private Long customerId;
 
     @Column(name = "name")
     private String name;
